Revert "Move more internal packages to be internal (#5700)"
This reverts commit e184d27b45ed586f9339070b74962ef43c61ef59.
